Split GUI config source tests
